fix(app-server): avoid false pong timeouts after stalls
