+++ src/websocket/listener/heartbeat.ts
@@ -1,3 +1,4 @@
+import { createMissedPongWatchdog } from "@/websocket/missed-pong-watchdog";
 import {
   LISTENER_HEARTBEAT_INTERVAL_MS,
   LISTENER_PONG_TIMEOUT_MS,
@@ -5,45 +6,8 @@ import {
 import { getListenerTransportKind, type ListenerTransport } from "./transport";
 import type { ListenerRuntime } from "./types";
 
-export interface MissedPongWatchdog {
-  shouldTerminate(lastPongAt: number | null): boolean;
-  recordPing(sentAt: number): void;
-}
-
 export interface ConnectionHeartbeatOptions {
   intervalMs?: number;
-}
-
-/**
- * Count actual unanswered heartbeat probes instead of elapsed wall time.
- *
- * A wall-clock-only watchdog cannot distinguish a dead peer from a locally
- * starved event loop. If the interval callback itself is delayed beyond the
- * timeout (CPU saturation, sleep/wake), it otherwise kills a healthy socket
- * before giving the peer a fresh ping to answer.
- */
-export function createMissedPongWatchdog(
-  maxUnansweredPings: number,
-): MissedPongWatchdog {
-  let lastPingAt: number | null = null;
-  let unansweredPings = 0;
-
-  return {
-    shouldTerminate(lastPongAt) {
-      if (
-        lastPingAt !== null &&
-        lastPongAt !== null &&
-        lastPongAt >= lastPingAt
-      ) {
-        unansweredPings = 0;
-      }
-      return unansweredPings >= maxUnansweredPings;
-    },
-    recordPing(sentAt) {
-      lastPingAt = sentAt;
-      unansweredPings += 1;
-    },
-  };
 }
 
 function getCurrentStreamTransport(
