+++ src/websocket/missed-pong-watchdog.ts
@@ -1,0 +1,36 @@
+export interface MissedPongWatchdog {
+  shouldTerminate(lastPongAt: number | null): boolean;
+  recordPing(sentAt: number): void;
+}
+
+/**
+ * Count actual unanswered heartbeat probes instead of elapsed wall time.
+ *
+ * A wall-clock-only watchdog cannot distinguish a dead peer from a locally
+ * starved event loop. If the interval callback itself is delayed beyond the
+ * timeout (CPU saturation, sleep/wake), it otherwise kills a healthy socket
+ * before giving the peer a fresh ping to answer.
+ */
+export function createMissedPongWatchdog(
+  maxUnansweredPings: number,
+): MissedPongWatchdog {
+  let lastPingAt: number | null = null;
+  let unansweredPings = 0;
+
+  return {
+    shouldTerminate(lastPongAt) {
+      if (
+        lastPingAt !== null &&
+        lastPongAt !== null &&
+        lastPongAt >= lastPingAt
+      ) {
+        unansweredPings = 0;
+      }
+      return unansweredPings >= maxUnansweredPings;
+    },
+    recordPing(sentAt) {
+      lastPingAt = sentAt;
+      unansweredPings += 1;
+    },
+  };
+}
